# docs/spec/api/openapi.yaml
openapi: 3.0.3
info:
  title: Enterprise IDP API
  description: >
    API for onboarding business processes, discovering available Azure AI
    Content Understanding analyzers, and triggering/reviewing document
    inference pipeline runs.


    This document is hand-written and authoritative: the FastAPI backend
    implements it, and the TypeScript client is generated from it.


    Authentication supports three caller modes: the `httpOnly`
    `cl_idp_session` browser cookie set by `POST /auth/login`, short-lived
    per-user bearer tokens returned by `POST /auth/token`, and a shared
    `SERVICE_API_TOKEN` bearer token for trusted workers and scripts.
  version: "0.2.0"
servers:
  - url: /api
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: Start a browser session
      security: []
      description: >
        Validates an end user's email/password, returns their public
        profile, and sets the `httpOnly` `cl_idp_session` cookie used by
        first-party browser calls. Unlike `POST /auth/token`, the JWT is
        not returned in the response body.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Signed-in user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/token:
    post:
      tags: [auth]
      summary: Mint a user bearer token
      security: []
      description: >
        Validates an end user's email/password and returns a short-lived
        `Authorization: Bearer` token for scripts, CLI calls, and ad-hoc
        testing. No cookie is set; the token carries the caller's own
        role, so RBAC stays the same as their browser session.
      operationId: issueToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Short-lived bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /processes:
    get:
      tags: [processes]
      summary: List business processes
      description: >
        Returns all business processes, sorted by `createdAt` descending
        (newest first). No pagination in this demo.
      operationId: listProcesses
      responses:
        "200":
          description: List of business processes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BusinessProcess"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [processes]
      summary: Create a business process
      description: >
        Creates the process and asynchronously provisions a Content
        Understanding routing analyzer over the selected analyzers (see
        `routingAnalyzerStatus` on the returned resource). Requires
        `IT Admin` for user-bound callers; trusted service-token
        automation is also allowed.
      operationId: createProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProcessInput"
      responses:
        "201":
          description: Created business process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProcess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: >
            A business process with this name already exists (case-
            insensitive match).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: >
            Content Understanding was unreachable while resolving analyzer
            display names, so the process was not saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /processes/{processId}:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [processes]
      summary: Get a business process
      operationId: getProcess
      responses:
        "200":
          description: Business process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProcess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [processes]
      summary: Update a business process
      description: >
        Overwrites the existing configuration (no versioning). The routing
        analyzer is rebuilt (resetting `routingAnalyzerStatus` to
        `building`) **only if `allowedAnalyzerIds` changed** — editing other
        fields leaves the process immediately usable.


        No optimistic concurrency: concurrent updates are last-write-wins.
        Requires `IT Admin` for user-bound callers; trusted service-token
        automation is also allowed.
      operationId: updateProcess
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessProcessInput"
      responses:
        "200":
          description: Updated business process
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProcess"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: >
            Another business process already has this name (case-
            insensitive match).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: >
            Content Understanding was unreachable while resolving analyzer
            display names, so the process was not saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [processes]
      summary: Delete a business process
      description: >
        Deletes the process, all of its jobs, and all of its stored
        documents. Irreversible, with no confirmation beyond the client's
        own.


        The Content Understanding analyzers provisioned for this process are
        intentionally **not** deleted; they are left in the CU account and
        identifiable by their `idp_` ID prefix. Requires `IT Admin` for
        user-bound callers; trusted service-token automation is also
        allowed.
      operationId: deleteProcess
      responses:
        "204":
          description: Process, jobs, and documents deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /analyzers:
    get:
      tags: [analyzers]
      summary: List available Content Understanding analyzers
      description: >
        Returns the analyzers selectable as a process's allowed forms: a
        curated catalog of prebuilt analyzers plus the custom analyzers in
        the configured Content Understanding account.


        Analyzers created and owned by this application (routing and derived
        analyzers, prefixed `idp_`) are excluded — they are internal
        plumbing, not user choices.
      operationId: listAnalyzers
      responses:
        "200":
          description: List of analyzers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Analyzer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          description: Content Understanding is unreachable or returned an error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
    ProcessId:
      name: processId
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Send either a short-lived user token minted by `POST /auth/token`
        or the shared `SERVICE_API_TOKEN` as `Authorization: Bearer
        <token>`.
    cookieAuth:
      type: apiKey
      in: cookie
      name: cl_idp_session
      description: >
        `httpOnly` browser session cookie set by `POST /auth/login`.
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: The request was invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or the supplied credentials were invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated user's role is not allowed to perform this action.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [code, message]
      description: >
        Uniform error body returned with every 4xx and 5xx response, with
        the single exception of `/healthz`, which returns a `Health` body on
        503 so that probes get dependency detail rather than a generic
        error.
      properties:
        code:
          type: string
          description: >
            Stable machine-readable identifier, e.g. `process_not_found`,
            `duplicate_process_name`, `unsupported_file_type`,
            `file_too_large`, `too_many_pages`, `routing_analyzer_not_ready`,
            `unknown_field_path`, `job_not_failed`,
            `content_understanding_unavailable`.
          example: unsupported_file_type
        message:
          type: string
          description: Human-readable explanation, safe to show to a user.
        details:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Optional structured context (e.g. the offending field path, the
            actual file size, the conflicting process ID).
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    TokenResponse:
      type: object
      required: [accessToken, expiresIn]
      description: >
        Response body for `POST /auth/token` — a short-lived bearer token.
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
          default: Bearer
        expiresIn:
          type: integer
          description: Seconds until `accessToken` expires.
    UserRole:
      type: string
      enum: ["IT Admin", Reviewer, "End User"]
      description: >
        Closed set of roles RBAC enforcement is based on.
    User:
      type: object
      required: [id, email, displayName, roleLabel, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        roleLabel:
          $ref: "#/components/schemas/UserRole"
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    BusinessProcessInput:
      type: object
      required:
        - name
        - description
        - allowedAnalyzerIds
        - confidenceThreshold
        - ownerEmail
      properties:
        name:
          type: string
        description:
          type: string
        allowedAnalyzerIds:
          type: array
          minItems: 1
          maxItems: 199
          uniqueItems: true
          items:
            type: string
          description: >
            IDs of selected Content Understanding analyzers. At most 199:
            Content Understanding permits 200 classification categories and
            the app reserves one for the `other` catch-all. The literal
            value `other` is reserved and rejected.
        confidenceThreshold:
          type: number
          format: float
          minimum: 0
          maximum: 1
          description: >
            Stored/transmitted as a 0–1 fraction. The UI collects/displays
            this as a 0–100% value and converts to/from this range.
        ownerEmail:
          type: string
          format: email
    BusinessProcess:
      allOf:
        - $ref: "#/components/schemas/BusinessProcessInput"
        - type: object
          required: [id, allowedAnalyzers, routingAnalyzerStatus, createdAt, updatedAt]
          properties:
            id:
              type: string
            allowedAnalyzers:
              type: array
              items:
                $ref: "#/components/schemas/AnalyzerRef"
              description: >
                Display names for `allowedAnalyzerIds`, resolved from the
                analyzer catalog at save time and stored with the process,
                so list screens can render labels without a live Content
                Understanding call. Presentation only —
                `allowedAnalyzerIds` remains authoritative, and a name may
                be stale if the analyzer was renamed since the last save.
            routingAnalyzerStatus:
              type: string
              enum: [building, ready, failed]
              description: >
                Provisioning status of the Content Understanding routing
                analyzer built over `allowedAnalyzerIds`. Rebuilt on create,
                and on update only when `allowedAnalyzerIds` changes.
                Uploads are rejected with 409 unless this is `ready`.
            routingAnalyzerError:
              type: string
              nullable: true
              description: >
                Human-readable failure reason, set when
                routingAnalyzerStatus is `failed` (including the case where
                a selected analyzer no longer exists in the Content
                Understanding account). Cleared when the next rebuild
                starts, so a `building` process never shows a stale error.
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    AnalyzerRef:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
        name:
          type: string
...